C++ LeetCode 150 Interview questions.

+++ src/CPP/code/leetcode/topInterview150/binary_tree_bfs/AverageOfLevelsInBinaryTree.cpp
@@ -110,7 +110,7 @@ public:
             queue<AverageOfLevelsInBinaryTreeNode *> levelB;
             levelA.push(root);
             long long int sum;
-            double count;
+            int count;
             AverageOfLevelsInBinaryTreeNode *processing;
             while (!levelA.empty() || !levelB.empty()) {
                 sum = 0;
@@ -128,7 +128,7 @@ public:
                             levelB.push(processing->left);
                         }
                     }
-                    result.push_back(sum / count);
+                    result.push_back((sum / (double) count));
                 } else {
                     while (!levelB.empty()) {
                         processing = levelB.front();
@@ -142,7 +142,7 @@ public:
                             levelA.push(processing->left);
                         }
                     }
-                    result.push_back(sum / count);
+                    result.push_back(sum / (double) count);
                 }
             }
         }
